Fuck Yeah! Got optim to work, and holy crap it solves the logistic regression incredibly way faster than using vanilla gradient descent.
I still need to figure out the variable scoping thing, as functions clearly have
access to values not explicitly passed in, which is different than what I'm used
to.

This makes a lot of sense as to why in Andrew Ng's class, we do not implement
gradient descent for logistic regression as it is very slow, for this particular
problem it required more than 1M iterations and still hadn't converged.