First try at one vs all classification
Unrelated note: fminunc uses an unconstrainted optimization algorithm, not
conjugate gradient. This is a line of investigation to look at for the logistic
regression implementation I wrote in R, because in R I was using conjugate
gradient with optim() function. Let's try again with an unconstrained algo and
see what happens.